+++ example/android/app/src/main/kotlin/dev/binarybush/tawk_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package dev.binarybush.tawk_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()