+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# toggle_list
+# tawk
 
 All notable changes to this project will be documented in this file.